fix: slim WhatsApp bridge image
