+++ docker/Dockerfile.whatsapp-bridge
@@ -17,9 +17,6 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     make \
     g++ \
     git \
-    libsqlite3-dev \
-    libsecret-1-dev \
-    pkg-config \
     && rm -rf /var/lib/apt/lists/*
 
 RUN npm install -g pnpm@9.15.0
@@ -31,8 +28,21 @@ RUN pnpm install --ignore-scripts
 
 COPY . .
 RUN pnpm run build
-RUN pnpm rebuild
-RUN pnpm prune --prod --ignore-scripts
+RUN mkdir -p /bridge-dist/crypto /bridge-dist/whatsapp-bridge \
+    && cp dist/crypto/canonical-hash.js /bridge-dist/crypto/ \
+    && cp dist/whatsapp-bridge/*.js /bridge-dist/whatsapp-bridge/
+
+RUN node -e "const fs=require('node:fs'); const root=JSON.parse(fs.readFileSync('package.json','utf8')); const wanted=['@whiskeysockets/baileys','pino','qrcode-terminal']; const dependencies=Object.fromEntries(wanted.map((name)=>[name,root.dependencies[name]])); console.log(JSON.stringify({type:root.type,packageManager:root.packageManager,dependencies},null,2));" > /tmp/whatsapp-bridge-package.json
+
+RUN mkdir -p /bridge-runtime \
+    && cp pnpm-lock.yaml .npmrc /tmp/whatsapp-bridge-package.json /bridge-runtime/ \
+    && mv /bridge-runtime/whatsapp-bridge-package.json /bridge-runtime/package.json \
+    && cd /bridge-runtime \
+    && pnpm install --prod --ignore-scripts --prefer-offline --no-frozen-lockfile
+
+RUN cd /bridge-runtime \
+    && pnpm rebuild \
+    && node -e "await import('@whiskeysockets/baileys'); await import('pino'); await import('qrcode-terminal'); await import('/bridge-dist/whatsapp-bridge/contract.js');"
 
 FROM node:22-bookworm-slim AS runtime
 
@@ -45,7 +55,6 @@ ARG USER_GID=1000
 
 RUN apt-get update && apt-get install -y --no-install-recommends \
     ca-certificates \
-    curl \
     && rm -rf /var/lib/apt/lists/* \
     && groupmod -g ${USER_GID} node \
     && usermod -u ${USER_UID} -g ${USER_GID} node \
@@ -53,9 +62,9 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     && chown -R node:node /home/node /data
 
 WORKDIR /app
-COPY --from=builder --chown=node:node /build/dist ./dist
-COPY --from=builder --chown=node:node /build/node_modules ./node_modules
-COPY --from=builder --chown=node:node /build/package.json ./
+COPY --from=builder --chown=node:node /bridge-dist ./dist
+COPY --from=builder --chown=node:node /bridge-runtime/node_modules ./node_modules
+COPY --from=builder --chown=node:node /bridge-runtime/package.json ./
 
 ENV NODE_ENV=production
 ENV WHATSAPP_BRIDGE_PORT=3004
